requirements.txt
python-fasthtml
uvicorn
claudette
python-dotenv
ell-ai